+++ Theme/resources/views/partials/header-single-sampyoGroup.blade.php
@@ -1,4 +1,4 @@
-<div class="overflow-x-hidden header-archive-root2">
+<div class="labtop:overflow-x-hidden header-archive-root2">
   <div class="labtop:overflow-hidden header-archive-root">
     <header class="header-archive">
       <div class="container xl:max-w-screen-xl mx-auto max:labtop:px-0">
@@ -41,7 +41,11 @@
     </header>
   </div>
 </div>
-<nav class="items-fixed hidden labtop:block">
+<nav
+  id="fixedMenu"
+  class="items-fixed hidden labtop:block"
+  :class="{'absolute bottom-72 footerFixed': footerShown}"
+>
   <ul class="items-fixed-wrap">
     @foreach ($fixedMenu as $item)
       <li class="items-fixed-item">